STCOR-616 provide useQueryLimit hook
The `useQueryLimit()` hook returns a promise that resolves to a number
which can be used in a query's limit clause when retrieving entries from
a lookup table and you want _all_ the records.

```
const limit = await useQueryLimit();
```

`useQueryLimit()` first looks for a tenant-level mod-settings entry, then
for `stripes.config.maxUnpagedResourceCount` from `stripes.config.js`,
and then finally gives up and returns the value of the constant
`MAX_UNPAGED_RESOURCE_COUNT`, currently set to 2000.

`usetTenantPreferences` was shamlessly copied from `usePreferences`;
both should be rewritten to share common code. Probably we want
`usePreferences` to be the top-level, which looks up a value in
user-settings, then tenant-settings.

Refs STCOR-616, STCOR-617