STCOR-899 show user-friendly labels on ECS pre-login screen (#1614)
The pre-login screen for ECS tenants with per-tenant authentication
configured shows a select-list with each tenant as an option. This PR
adds the ability to read a new attribute, `displayName`, from the
tenant-object to be used as the label on that list in order to make it
more user friendly.

Previously, the list in `stripes.config.js::config.tenantOptions` was
shaped like this:
```
{
  k1: { name: "k1", clientId: "k1-application" },
  k2: { name: "k2", clientId: "k2-application" },
}
```
but now there is an optional, additional key, so it can be shaped like
```
{
  k1: { name: "k1", clientId: "k1-application", displayName: "Milli" },
  k2: { name: "k2", clientId: "k2-application", displayName: "Vanilli" },
}
```
This change should be 100% backwards compatible. The tenant value
remains duplicated across the key and the `name` attribute, and the
select-list will fall back on `name` if `displayName` is not present,
preserving existing behavior.

Refs STCOR-899

(cherry picked from commit 7d490e266d2083fbe70421c930fefc27bffaa510)