STCOR-964: Sort app navigation by localized displayName by default. (#1617)
Currently, apps order defaults to the order from stripes-config. This is
not-as-nice for users when they reset defaults and expect to have a
predictable time finding their app's name in the list so that they can
move it to it's preferred position.

This PR sorts them by localized `displayName` to save their eyes.

(cherry picked from commit 165587ccbf82dfb1442759f968c9424ca5c3b186)