STCOR-968: Enhance locale management to support user and tenant locale configurations from both mod-configuration and mod-settings. (#1628)
## Description
A new setting in `My profile` -> `Language & Localization` is created,
which should use mod-settings API instead of deprecated
mod-configuration. It was decided not to make a breaking change for
stripes-core and still support mod-configuration in a
backwards-compatible way.

The main changes are in the `loadResources` function. First, if the user
can read mod-settings, it fetches locale information using two endpoints
(one for tenant and one for the user). When these calls return, it
checks if at least one response contains a valid setting. If so, it
applies those settings; otherwise, if the user has mod-configuration
reading permission, it fetches locale information using the
mod-configuration API for both user and tenant settings and applies
them. If the user doesn't have mod-settings permission but has
mod-configuration permission, it fetches locale information using the
mod-configuration API. Additionally, if the user has mod-configuration
permission, it fetches plugins and bindings. It also checks if the okapi
instance is not in "withoutOkapi" mode, it fetches module information
using `discoverServices`. Finally, it returns a flattened array of all
the results from the promises.

## Changes in other modules
https://github.com/folio-org/ui-myprofile/pull/251
https://github.com/folio-org/ui-tenant-settings/pull/456
https://github.com/folio-org/ui-developer/pull/468
https://github.com/folio-org/stripes-smart-components/pull/1577
https://github.com/folio-org/ui-plugin-query-builder/pull/246


## Issues
[STCOR-968](https://folio-org.atlassian.net/browse/STCOR-968)

## Screencasts

https://github.com/user-attachments/assets/d1131450-6d43-4e6d-91e8-71d4d0633616