STCOR-970 validateUser should return null on failure (#1629)
`validateUser()` is expected to return session data on success or null
on failure. Previously, it wrongly returned an error message on failure,
causing its calling function (`checkOkapiSession()`) to act as though an
existing session were being restored (which of course is not the case
given that session restoration had just failed...) and skip the
SSO-check that happens only once, when stripes is initializing.

Related, restore the CSS for the "Log in via SSO" button so it remains
prominently visible when it is present.

Approach

If cookies are revoked for any reason -- maybe a user deletes them,
maybe the backend revokes them -- then the UI will be in a state where
it has valid session data in redux and in localstorage, but it no longer
has the corresponding cookies. This is an awkward state for Stripes,
because it doesn't know it's in this state, because the cookies are
HTTPOnly, so it never knows if cookies are acutally present.

When reloading, the valid-session/invalid-cookies state was particularly
problematic. Stripes would find session data in storage and thus try to
restore the session with a request to `.../_self`, a request which would
fail due to the missing cookies. Because `validateUser()` was
incorrectly returning an error message, however, this caused Stripes to
skip the SSO-check that normally runs when Stripes is initializing
without an active session.

Minor refactoring of `checkOkapiSession()` and `validateUser()` (the
former now passes an error-handler to the latter, because DI simplifies
testing) is a happy side-effect. The actual bugs being squashed are:

* `validateUser()` now returns undefined on error, and correctly calls an
  error-handler.
* "Log in via SSO" button CSS is restored.

Refs STCOR-970

(cherry picked from commit 81ce56c02770515398235c19f4cf510fed24e42f)