STCOR-972 passing tenant as a URL param to keycloak redirect_uri (#1634)
## Description
I'm going to use @zburke explanation on this because I couldn't have
explained the issue better myself

> In an ECS setup, there are three required steps to authentication;
keycloak is step-2, and skipping step-1 causes step-3 to fail:
> 1. choose a tenant and save it to local-storage, then redirect to
keycloak
> 2. authenticate in keycloak, then redirect to stripes with an OTP in
the URL
> 3. make an API request to trade the OTP for AT/RT cookies,
/authn/token?otp=${OTP from step-2} with the X-Okapi-Tenant header set
to the tenant from step-1
>
> Bookmarking the authentication URL in step-2 means the tenant is never
stored to local-storage, thus causing step-3 to fail.
IIRC, we can pass values to Keycloak in the URL and it will include them
along with the OTP when it redirects back to Stripes.

Instead of storing selected `tenant` and `client_id` in `localStorage`,
we can in step 1 pass them as part of `redirect_uri`, so that we can
read them from url and use in step 3.
We're still keeping code related to storing tenant in localStorage,
because we actually need to keep the tenant that a user has initially
logged in with, so if they change affiliation - we can logout with the
initial tenant.

## Screenshots


https://github.com/user-attachments/assets/6c031f1d-5659-4049-99fc-419af59bbce8



## Issues
[STCOR-972](https://folio-org.atlassian.net/browse/STCOR-972)