STCOR-978 Use BroadcastChannel to isolate token to last-focused window. (#1647)
Refs [STCOR-978](https://folio-org.atlassian.net/browse/STCOR-978)
(Sunflower backport:
[STCOR-84](https://folio-org.atlassian.net/browse/STCOR-984))

This PR resolves an issue where users experienced an abrupt logout due
to failures in token rotation across multiple tabs of usage - one bad
apple among all of their open tabs spoiled it for the rest.

Only a single, active tab is able to actually rotate. The rest just
schedule their next rotation and *don't* actually rotate. This will be
the last tab that the user interacted with and will persist through
usage of other applications.

With this approach, each new tab (single `FFetch` instance) creates a
unique id for the tab and sets up a `focusin` handler so that when a new
tab is focused, it will broadcast the focused tab's unique id so that
other windows can set it in their session and use it to check whether
token rotation should occur. If the id in session doesn't match the
current tab's id, rotation will not occur in that tab. All tabs share
the same cookie, so as long as its kept alive by the current window, it
will remain for other tabs to use. Each tab stores the last-focused tab
in their session.

Other attempts:
- relying solely on `document.hasFocus()` to dictate whether or not a
window *can* rotate. This is limited because the browser tab that only
has the tab focus (and focus isn't in the UI) will not be considered
focused, leaving the status in limbo. It also causes rtr to not happen
if other apps are in use when rotation happens - and we don't need that!

- syncing via `storage` event. This seemed very difficult to debug - the
events didn't seem to fire with sessionStorage updates across windows. I
ended up switching to a `localStorage` implementation, but even that was
problematic. The events just didn't want to stick across windows. The
persistent nature of `localStorage` disqualifies it as a lone
possibility since the previously generated window id would stick around
and fail to match a new window after the entire application is closed.

The video displays the logic in action (with rtr logging set in
stripes-config) you can see when the focus changes and the newly focused
tab's id is broadcast to all of the other tabs...


https://github.com/user-attachments/assets/fb1fffee-c698-4c56-9904-9608bc955f5f
(cherry picked from commit 9676cbc8d52ddb63afe822beb1eb22efb47874cf)

This PR appends the focus handler to check the access token expiry and
possibly refresh before any data calls can be made.
This resolves a situation where a user navigates away from FOLIO
entirely (different window, app, etc) and closes their most recent
window without going to any previously used FOLIO tab... once they
refocus a FOLIO tab, if the at is expired, the logic will refresh.

(cherry picked from commit e739f7b78e68406ea922c45d7cb5f87e53816642)

---------

Co-authored-by: John Coburn <jcoburn@ebsco.com>