STCOR-986 handle OTP in StrictMode (#1652)
Effects run twice in `<StrictMode>`, but OTP must only happen once; make
it so.

This PR changes the way OIDCLanding handles OTP exchange, moving it from
an effect (which will run x2 in development when `<StrictMode>` is
turned on, which of course we want it to be), outside the component
lifecycle so it only runs once no matter how many times the component is
mounted/unmounted/remounted in development-mode.

It may bother you that there is both an event-handler and an effect to
handle the same thing, and that's fair. The problem is, there is a race
condition between the OTP-exchange request and the component loading,
and we want to initialize the session no matter who wins the race:

* if the component loads before the request finishes, then storage will
  be empty when the effect runs so nothing will happen then 😢 but the
  event handler will catch the event emitted when the request finally
  finishes 😍
* if the request finishes before the component loads, then there won’t
  be a handler to listen for the event so nothing will happen 😢 but the
  effect runs when the component mounts and will find the data in storage
  😍

Refs STCOR-986