Revert "STCOR-986 handle OTP in StrictMode" (#1656)
Reverts folio-org/stripes-core#1652

Instead of introducing a race condition between the OTP-exchange request
and component init and then very carefully handling the race condition,
a better approach will not introduce the race condition at all, and will
not call `otpExchange()` at the module level.

This module-level call is particularly gross, a definite anti-pattern
that raised eyebrows in the original PR but which I waved away. This
call appears to cause particular grief in the end-to-end test
environment, which is to say, my attempt to reliably handle the race
condition failed.

A better approach to application init and OTP exchange will avoid the
module-level call. This will avoid the race condition, and avoid
invoking this code path except when explicitly called for by accessing
the `/oidc-landing` route.