STCOR-987 handle forgot-username, forgot-password for ECS tenants
The only business-logic change here is the insertion of
`<SelectAndDispatchTenant />` in the forgot-* forms, and configuring
their submit buttons to be disabled when a tenant cannot be determined.

* Single-tenant: `<SelectAndDispachTenant />` will return null because
  the tenant can be gleaned from `stripes.okapi`.
* ECS: `<SelectAndDispatchTenant />` will render a `<Select />` menu
  with an onChange handler that dispatches the selection (causing it to
  be populated in `stripes.okapi`), allowing the submit button to be
  enabled.

Previously, `stripes.okapi.tenant` was unset in ECS mode prior to
authenticating (see also STCOR-985 / PR #1649 for single-tenant mode),
leading to an NPE when these components submitted and ran API queries.

This PR looks bigger than it is because it cleans up a lot of cruft. To
wit:

* queries are rewritten from stripes-connect -> react-query
* redux interactions are mostly gone (they were carefully added in
  `mapStateToProps` and then cleared on unmount. Um, wha...? This also
  means we can remove the wrapper component entirely. Redux remains only
  in `<SelectAndDispatchTenant />`.
* Directories' default exports match the directory names.
* Tests are a thing.

Refs STCOR-987