STCOR-987 handle forgot-username, forgot-password for ECS tenants (#1657)
The only business-logic change here is the insertion of
`<SelectAndDispatchTenant />` in the forgot-* forms, and configuring
their submit buttons to be disabled when a tenant cannot be determined.

* Single-tenant: `<SelectAndDispachTenant />` will return null because
  the tenant can be gleaned from `stripes.okapi`.
* ECS: `<SelectAndDispatchTenant />` will render a `<Select />` menu
  with an onChange handler that dispatches the selection (causing it to be
  populated in `stripes.okapi`), allowing the submit button to be enabled.

Previously, `stripes.okapi.tenant` was unset in ECS mode prior to
authenticating (see also STCOR-985 / PR #1649
for single-tenant mode), leading to an NPE when these components
submitted and ran API queries.

This PR looks bigger than it is because it cleans up a lot of cruft. To
wit:

* queries are rewritten from stripes-connect -> react-query.
* redux interactions are mostly gone (previously, they were carefully
  added in `mapStateToProps` and then cleared on unmount. Um, wha...?)
  This also means we can remove the wrapper component entirely. Redux
  remains only in `<SelectAndDispatchTenant />`.
* directories' default exports match the directory names.
* tests are rewritten from BigTest -> Jest.
* omit `Login/LoginForm.js`, which appears to have been inadvertently
  introduced during keycloakification in PR #1532
  (2bea8b0674497961635431fbc7d7e260244afd8e).
* omit duplicate CSS files, which SonarCloud would do well to notice
  when whining about duplicate code 😒. Let's not allow the perfect to be
  the enemy of the good, hey?

Refs STCOR-987

(cherry picked from commit e3c1d9409bac5d11ced126e4b0e15257c0431766)