STCOR-989: Generalise useChunkedCQLFetch hook to allow for other chunked id transformations (#1653)
## feat: useChunkedIdTransformFetch hook

### Current state
useChunkedCQLFetch is a pre-existing hook which aims to allow for
frontend joins, taking in some list of UUIDs and performing a fetch for
those objects in a chunked and reliable manner.

### Premise
This hook, as its name suggests, is currently only applicable to
endpoints making use of CQL, where the bulk of the internal logic for
actually chunking said requests is API agnostic.

Hence this PR aims to generalise that logic, leaving the API for
useChunkedCQLFetch in place to avoid backwards compatibility issues, but
allowing for frontends pointing at services using other API shapes to
make use of the same underlying chunked-fetch logic.


### Use case
In particular, ui-agreements has an upcoming use case where a list of
IDs is presented from an external service, and we have to check whether
those match anything in our internal DB. The API for this already
exists, but the list of IDs can become very long, and the query requires
filtering on a couple of fields, so the query string grows very quickly,
and regularly returns `414 URI Too Long`.

In particular  the query string is something like:
```?filters=((pci.identifiers.identifier.status == approved && pci.identifiers.identifier.namespace == 'gokb_uuid' && pci.identifiers.identifier == ID_1) || (... ID_2) || (... ID_3))```

### Approach
This PR moves the bulk of the logic which previously lived inside useChunkedCQLFetch up to a new generic hook: `useChunkedIdTransformFetch`, and keeps all CQL-specific logic confined to the old hook (such as the presence of the `query` and `limit` parameters, as well as the specific way to map a list of Ids into such a query)


### Considerations
There is absolutely a case here to suggest that part of the problem is that there ought to be a backend API to handle the namespacing and approved status matching parts, but fundamentally we have an unknown size list of ids coming in, and unless we include those in a HTTP body, the risk of seeing a 414 is there, and the argument about whether it's acceptable to fetch records via some kind of `POST` with a body is a different kettle of fish.