STCOR-977: Listen for SessionReady events and update state to trigger re-render. (#1667)
## Description
> **Summary**: Listen for a `SessionReady` event in `RootWithIntl` and
re-render when it is detected (as provided by `@folio/stripes-loader` in
`mod-fedified-stripes`), in addition to re-rendering when
`props.isAuthenticated` changes (as provided by redux in `stripes-core`
in legacy-stripes).

> **Details**: add an event handler that leverages `setState()`, i.e.
when it detects a given event it calls a `setState()` based-function,
and that piece of state (along side `props.isAuthenticated`) will cause
the component to re-render.

## Issues
[STCOR-977](https://folio-org.atlassian.net/browse/STCOR-977)