STCOR-1000: Take `tenant` from URL for multi-tenant setup in `getLoginTenant`. Make `makeCall` method async to avoid displaying an error message when loading the password reset page. (#1670)
Multi-tenant setup.
It can be reproduced in Sunflower ECS.

When entering a new password on the `/reset-password` page, the password
requirements should appear under the field.

The requirements are not displayed because the `/rules` request fails
due to the absence of a tenant. The tenant could be taken from the URL,
but this doesn't happen bacause the client id is missing, and the
`(tenant && clientId)` condition prevents the tenant from being returned
from `getLoginTenant`.

Remove the `(tenant && clientId)` conditions and derive tenant and
clientId from URL and from the next source (tenantOptions/stripesOkapi)
only if they are empty.

Fixed an issue where the `System error. Please wait a few minutes and
try again.` message would momentarily appear when loading the password
reset page. The fix was to return a Promise in the `makeCall` method.

[STCOR-1000](https://folio-org.atlassian.net/browse/STCOR-1000)

Recording of the bug

https://github.com/user-attachments/assets/8251e97c-8b2b-4d88-a546-8aa5dec9b6cb

Recording after the fix is applied (without the fix of `makeCall`)

https://github.com/user-attachments/assets/9bb2a94f-3ac0-4c03-b3a7-e1e03666babf
(cherry picked from commit 714559e0618b649dfcbbba17fd52e46c6af2f940)