STCOR-805 supply `modules` on Stripes
Supply `modules` to Stripes when initializing it, and for each module
supply a module's `name` attribute from its `package.json` file in the
`data` argument passed to an event handler.

`modules` is an object grouped by `type` typically shaped like
```
   app: [{ module: '@folio/users' }, { module: '@folio/checkin' }],
   settings: [{ module: '@folio/users' }, { module: '@folio/checkin' }],
   plugin: [ ... ],
   handler: [ ... ],
```
but it could also have a key like
```
  servicepointsConsumer: [{ module: '@folio/checkin' }]
```
that stripes-core does not know or care about. A handler-module like
ui-servicepoints could leverage that key in a `SELECT_MODULE` handler to
see if a value in the `stripes.modules.servicepointsConsumer` array
matches the value from `data.module` (the incoming module). This allows
modules to use their `actsAs` types to interact independently of any
particular knowledge of those types in stripes-core. Note that in the
example above, ui-checkin publishes a type and ui-servicepoints consumes
that type, and stripes-core is utterly uninvolved.

Refs STCOR-805