STCOR-635 preload modules in bundles built with --lazy
When a bundle is constructed with `--lazy()`, the module-loader function
will be `getDynamicModule()` implemented with `import()` instead of
`getModule()` implemented with `require()`, i.e. the module-loader
function will be async. That offers the opportunity to implement
lazy-loading, but for FOLIO this has caveats:
* event-handler modules provide static methods, meaning they must be
  accessible synchronously during render
* plugin modules may need to be loaded synchronously due to how they are
  loaded in the UI

IOW, module loading cannot always be handled with lazy. Here, we take
the simple approach: pre-load everything, which may seem to fly in the
face of implementing code-splitting to begin with, but the end-goal here
is not, in fact code-splitting with separate webpack chunks, but
code-splitting with independently constructed module-federated bundles.
In that context, being able to load separate pieces is still a win.

We may eventually choose a more elaborate process for loading modules,
e.g. choosing to lazy-load applications but to pre-load other types, but
we're starting with a simple approach: pre-load everything.

Refs STCOR-635