STCOR-1014 given a code-split bundle, load and cache all modules during getModules() (#1673)
Summary

Given a code-split bundle where modules are loaded dynamically, load and
cache all modules during initialization in `getModules()`

Details

In a monolithic bundle, stripes-config supplies synchronous
`getModule()` functions implemented with `require()`. In a code-split
bundle, stripes-config supplies asynchronous `getDynamicModule()`
functions implemented with `import()`. This PR adds handling for the
code-split bundle while remaining backwards compatible with monolithic
bundling.

The only real work in this PR is in `getModules()`, the async function
introduced in PR #1661 that converted the entire module-loading
infrastructure to async. We leverage this function to load and cache all
modules, creating a synchronous `getModule()` function analogous to the
one provided by stripes-config in a monolithic build, because
synchronous access is necessary when modules are accessed during
`render()` to invoke event-handlers and use plugins.

Renaming of things inside `handlerService.js` has not changed the
implementation at all; rather, the new names reflect what these
functions have always done: the handlers are invoked, possibly for their
side-effects or possibly for the components they return when they want
to interrupt the render process.

* See also https://github.com/folio-org/stripes-webpack/pull/171 This
  work is backwards compatible; it still supports non-lazy (monolithic)
  builds.

Refs STCOR-635, STRIPES-999, STCOR-1014