STCOR-1019 preserve existing public exports
Sigh; this package exports its API from `./index.js`, not
`./src/index.js`, which instead is the root of the application. I messed
that up in PR #1673. Someday, we'll separate the application-y bits of
this into stripes-ui and leave the shared components here, and use
consistent entry points in all our packages, and eliminate redux-form,
and win lotteries, and get puppies, and ponies, and ride unicorns, and
fart rainbows. Today though, today we'll just focus on avoiding
carelessly renaming exports.

Refs STCOR-1014, STCOR-1019